+++ lua/codecompanion/config.lua
@@ -150,7 +150,17 @@ local defaults = {
         opts = {
           auto_submit_errors = false, -- Send any errors to the LLM automatically?
           auto_submit_success = true, -- Send any successful output to the LLM automatically?
+          folds = {
+            enabled = true, -- Fold tool output in the buffer?
+            failure_words = { -- Words that indicate an error in the tool output. Used to apply failure highlighting
+              "error",
+              "failed",
+              "invalid",
+              "rejected",
+            },
+          },
           wait_timeout = 30000, -- How long to wait for user input before timing out (milliseconds)
+
           ---Tools and/or groups that are always loaded in a chat buffer
           ---@type string[]
           default_tools = {},
@@ -433,6 +443,12 @@ local defaults = {
           callback = "keymaps.goto_file_under_cursor",
           description = "Open the file under cursor in a new tab.",
         },
+        copilot_stats = {
+          modes = { n = "gS" },
+          index = 20,
+          callback = "keymaps.copilot_stats",
+          description = "Show Copilot usage statistics",
+        },
       },
       opts = {
         blank_prompt = "", -- The prompt to use when the user doesn't provide a prompt
@@ -523,10 +539,6 @@ local defaults = {
               context.filetype
             )
           end,
-          opts = {
-            visible = false,
-            tag = "system_tag",
-          },
         },
       },
     },
@@ -550,9 +562,6 @@ local defaults = {
                 context.filetype
               )
             end,
-            opts = {
-              visible = false,
-            },
           },
           {
             role = constants.USER_ROLE,
@@ -610,8 +619,8 @@ Your instructions here
 
 You are required to write code following the instructions provided above and test the correctness by running the designated test suite. Follow these steps exactly:
 
-1. Update the code in #buffer using the @insert_edit_into_file tool
-2. Then use the @cmd_runner tool to run the test suite with `<test_cmd>` (do this after you have updated the code)
+1. Update the code in #{buffer} using the @{insert_edit_into_file} tool
+2. Then use the @{cmd_runner} tool to run the test suite with `<test_cmd>` (do this after you have updated the code)
 3. Make sure you trigger both tools in the same response
 
 We'll repeat this cycle until the tests pass. Ensure no deviations from these steps.]]
@@ -660,9 +669,6 @@ We'll repeat this cycle until the tests pass. Ensure no deviations from these st
 3. Explain each function or significant block of code, including parameters and return values.
 4. Highlight any specific functions or methods used and their roles.
 5. Provide context on how the code fits into a larger application if applicable.]],
-          opts = {
-            visible = false,
-          },
         },
         {
           role = constants.USER_ROLE,
@@ -715,9 +721,6 @@ We'll repeat this cycle until the tests pass. Ensure no deviations from these st
       - Edge cases
       - Error handling (if applicable)
 6. Provide the generated unit tests in a clear and organized manner without additional explanations or chat.]],
-          opts = {
-            visible = false,
-          },
         },
         {
           role = constants.USER_ROLE,
@@ -775,9 +778,6 @@ Ensure the fixed code:
 - Is formatted correctly.
 
 Use Markdown formatting and include the programming language name at the start of the code block.]],
-          opts = {
-            visible = false,
-          },
         },
         {
           role = constants.USER_ROLE,
@@ -819,9 +819,6 @@ Use Markdown formatting and include the programming language name at the start o
         {
           role = constants.SYSTEM_ROLE,
           content = [[You are an expert coder and helpful assistant who can help debug code diagnostics, such as warning and error messages. When appropriate, give solutions with code snippets as fenced codeblocks with a language identifier to enable syntax highlighting.]],
-          opts = {
-            visible = false,
-          },
         },
         {
           role = constants.USER_ROLE,
@@ -935,7 +932,6 @@ This is the code, for context:
       prompts = {
         {
           role = constants.SYSTEM_ROLE,
-          opts = { visible = false },
           content = function()
             local schema = require("codecompanion").workspace_schema()
             return fmt(
@@ -978,7 +974,7 @@ You must create or modify a workspace file through a series of prompts over mult
 
             local ok, _ = pcall(require, "vectorcode")
             if ok then
-              prompt = prompt .. " Use the @vectorcode tool to help identify groupings of files"
+              prompt = prompt .. " Use the @{vectorcode_toolbox} tool to help identify groupings of files"
             end
             return prompt
           end,
@@ -1000,8 +996,10 @@ You must create or modify a workspace file through a series of prompts over mult
     },
     chat = {
       icons = {
-        pinned_buffer = " ",
-        watched_buffer = "󰂥 ",
+        buffer_pin = " ",
+        buffer_watch = "󰂥 ",
+        tool_success = "",
+        tool_failure = "",
       },
       debug_window = {
         ---@return number|fun(): number