+++ tests/config.lua
@@ -217,6 +217,14 @@ return {
         opts = {
           system_prompt = "My tool system prompt",
           wait_timeout = 3000,
+          folds = {
+            enabled = false,
+            failure_words = {
+              "error",
+              "failed",
+              "invalid",
+            },
+          },
         },
       },
       variables = {
@@ -364,8 +372,10 @@ return {
   display = {
     chat = {
       icons = {
-        pinned_buffer = " ",
-        watched_buffer = "👀 ",
+        buffer_pin = " ",
+        buffer_watch = "👀 ",
+        tool_success = "!!",
+        tool_failure = "xx",
       },
       show_references = true,
       show_settings = false,